`size_t`:remove

+++ include/cer0.h
@@ -1,8 +1,6 @@
 #ifndef cer0_h
 #define cer0_h
 
-#include <stdlib.h>
-
 #include "constants.h"
 #include "phase.h"
 #include "signals.h"
@@ -13,7 +11,7 @@ float *createNoteTable(
   float
 );
 
-size_t sizeofNoteTable(
+unsigned int sizeofNoteTable(
   unsigned char,
   unsigned char
 );
